구매확정 데이터를 조회하는 Reader만들기

+++ src/main/kotlin/com/settlement/settlementkotlin/domain/entity/Seller.kt
@@ -11,7 +11,7 @@ data class Seller (
     @Column(name = "seller_no")
     val id: Long,
     val sellerName: String,
-    val businessNo: Long,
+    val businessNo: Int,
     val sellType: String = "C",
     val bankType: String,
     val accountNo: Long,
